requirements.txt
PyQt6
pymongo
python-dotenv